Rewrite Stage 1: Rewrite the ray throwing stream
The Sound Physics rewrite is finally underway! I have rewritten much of `SoundPhysics.evaluateEnvironment`, most of the changes having to do with the Ray throwing stream. Assuming it works properly when this is all done, it should bring with it a much better performance profile, as well as more accurate reverb.

You'll notice that much of the math done in the original code has been removed, this is because i plan to pass the data generated in `evaluateEnvironment()` through to `processEnvironment`, and iterate over it again there. This serves many purposes, including better readability, intuitivity, and organization, but also better performance and accuracy:
- From what i have been reading lately, splitting up a pipeline into two paralell streams can sometimes help decrease thread parking time, which helps with performance.
- Having access to the complete dataset in `processEnvironment()` will allow me to more accurately process the individual ray data.
- I have eliminated the need for any objects with special concurrency code, while maintaining full concurrency, which should improve performance somewhat.
- I have also eliminated the need for any unnecesarily complex objects, by using simpler ones where possible, and creating my own where not. Notice the change from Lists/Maps to Sets, and the creation of the `com.thedocruby.soundphysics.RayResult` class to hold the ray data.

**WARNING:** This code is not complete nor tested, and despite how tempting it is, it should not be used in your game. I have not integrated the new ray data format with `processEnvironment()` to intentionally make it hard to successfully compile the mod here. Please DO NOT attempt to compile this code, and DO NOT ask me how to do it on the discord.

I plan to rewrite these areas of the code before I release 1.0.0-alpha.1. They are listed in no particurlar order:
- `com.thedocruby.soundphysics.SoundPhysics.processEnvironment()`
- `com.thedocruby.soundphysics.ALstuff.SPEfx`
- `com.thedocruby.soundphysics.performance.RaycastFix`
- The logging and debug features
- Much of the config and presets (this will be last before the alpha)

Once I have a mostly working alpha version of the rewrite, I will put a compiled JAR on my discord in the as of yet virgin #public-testing channel.