Workaround removal of ChunkSection#getYOffset()
```chunkSections[0].getYOffset()```
Formerly, when constructing a `ChunkSection` requires the int parameter `chunkPos`. The now removed field `yOffset` in the said class was calculated as `blockCoordFromChunkCoord(chunkPos)` or equivalently `chunkPos << 4`.

```chunkSections[0].param0$chunkPos << 4```
Time to address the parameter chunkPos. The mixin class `WorldChunkMixin` targets the Minecraft's class `Chunk` where its method `fillSectionArray` happens to be where `ChunkSection`s are constructed. Replace the value used there for `chunkPos`.

```-param0$world.sectionIndexToCoord(i) << 4```
The variable `i` from `fillSectionArray` is the index of the chunkSection. Wait, `world` is a parameter of `fillSectionArray`. This method is actually called from the constructor. Replace `world` with its value.

```-heightLimitView.sectionIndexToCoord(i) << 4```
`heighLimitView` is a parameter passed to the constructor _and_ also stored later as a field.